feat(#78): Progresso da Trilha
Alteração do modo de ingresso em jornadas
